Move KerrHorizon to GeneralRelativity
