+++ src/PointwiseFunctions/GeneralRelativity/KerrHorizon.hpp
