+++ tests/Unit/ParallelAlgorithms/ApparentHorizonFinder/CMakeLists.txt
@@ -30,6 +30,7 @@ target_link_libraries(
   PRIVATE
   ApparentHorizonFinder
   FiniteDifference
+  GeneralRelativity
   GeneralRelativitySolutions
   InterpolationHelpers
   LinearOperators
